refactor(debian): split flat debian.rs into directory module per ADR 0009
Split the 3,878-line debian.rs into a debian/ directory module with
7 focused sub-modules (utils, control, dsc, tarball, file_list,
copyright, deb) plus co-located test and golden test files. External
test files moved into the directory. All 107 tests pass unchanged.
